+++ NEWS.rst
@@ -7,6 +7,8 @@ every change, see the Git log.
 Latest
 ------
 * tbd
+* Major: Upgrade to kodo-python 13
+* Minor: Added an option to specify the finite field for the simulation.
 
 1.0.0
 -----